AI_core/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
pymongo==4.6.0
redis==5.0.1
weaviate-client==4.9.3
